+++ Runtime/EditorComment.cs
@@ -9,15 +9,24 @@ namespace OcUtility
     public class EditorComment : MonoBehaviour, IHierarchyIconDrawable
     {
 #if UNITY_EDITOR || DEBUG_BUILD
-        public string IconPath => "Assets/Oc Utility/Editor/Resources/EditorComment Icon.png";
+        public string IconPath => "EditorComment Icon";
         public int DistanceToText => -55;
-
+        [HideInInspector] GameObject gizmoTarget;
         public Context[] Contexts;
 
         void Reset()
         {
             Contexts = new Context[1];
+            gameObject.hideFlags = HideFlags.DontSaveInBuild;
         }
+
+        void OnDrawGizmosSelected()
+        {
+            Gizmos.color = ColorExtension.Rainbow();
+            Gizmos.DrawWireCube(transform.position, Vector3.one * 0.3f);
+            Gizmos.DrawWireSphere(transform.position, 0.1f);
+        }
+
 
         [Serializable]
         public class Context